implement navigation to search
