+++ core/model/src/main/java/com/dante/calorietracker/core/model/SearchArgs.kt
@@ -1,0 +1,8 @@
+package com.dante.calorietracker.core.model
+
+data class SearchArgs(
+    val mealType: String,
+    val dayOfMonth: Int,
+    val month: Int,
+    val year: Int
+)
